refactor: move pool-detail staking components
