+++ packages/web/src/views/pool/pool-detail/components/position-history-list/PositionHistoryList.styles.ts
