+++ packages/web/src/views/pool/pool-detail/components/position-history-list/position-history-table/position-info/PositionInfo.styles.ts
