+++ packages/web/src/views/pool/pool-detail/components/staking/Staking.styles.ts
