+++ packages/web/src/views/pool/pool-detail/components/staking/Staking.tsx
@@ -1,14 +1,20 @@
 import React from "react";
-import StakingContent from "@components/pool/staking-content/StakingContent";
-import StakingHeader from "@components/pool/staking-header/StakingHeader";
-import { StakingAnchor, StakingWrapper } from "./Staking.styles";
-import { DEVICE_TYPE } from "@styles/media";
-import { PoolPositionModel } from "@models/position/pool-position-model";
+
 import { PoolDetailModel } from "@models/pool/pool-detail-model";
+import { PoolStakingModel } from "@models/pool/pool-staking";
+import { PoolPositionModel } from "@models/position/pool-position-model";
+import { DEVICE_TYPE } from "@styles/media";
+
+import StakingContent from "./staking-content/StakingContent";
+import StakingHeader from "./staking-header/StakingHeader";
+
+import { StakingAnchor, StakingWrapper } from "./Staking.styles";
 
 interface StakingProps {
+  pool: PoolDetailModel | null;
   totalApr: string;
   stakedPosition: PoolPositionModel[];
+  poolStakings: PoolStakingModel[];
   breakpoint: DEVICE_TYPE;
   mobile: boolean;
   isDisabledButton: boolean;
@@ -16,13 +22,14 @@ interface StakingProps {
   handleClickStakeRedirect: () => void;
   handleClickUnStakeRedirect: () => void;
   loading: boolean;
-  pool: PoolDetailModel | null;
   isOtherPosition: boolean;
 }
 
 const Staking: React.FC<StakingProps> = ({
+  pool,
   totalApr,
   stakedPosition,
+  poolStakings,
   breakpoint,
   mobile,
   isDisabledButton,
@@ -30,7 +37,6 @@ const Staking: React.FC<StakingProps> = ({
   handleClickStakeRedirect,
   handleClickUnStakeRedirect,
   loading,
-  pool,
   isOtherPosition,
 }) => {
   return (
@@ -50,6 +56,7 @@ const Staking: React.FC<StakingProps> = ({
           pool={pool}
           totalApr={totalApr}
           stakedPosition={stakedPosition}
+          poolStakings={poolStakings}
           breakpoint={breakpoint}
           mobile={mobile}
           type={type}
