+++ packages/web/src/views/pool/pool-detail/components/staking/staking-content/StakingContent.styles.ts
