+++ packages/web/src/views/pool/pool-detail/components/staking/staking-content/incentivized-token-detail-tooltip-content/IncentivizeTokenDetailTooltipContent.styles.ts
