+++ packages/web/src/views/pool/pool-detail/components/staking/staking-content/staking-content-card/StakingContentCard.tsx
