+++ packages/web/src/views/pool/pool-detail/components/staking/staking-header/StakingHeader.stories.tsx
